pyproject.toml
[project]
name = "agent-class1"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "exa-py>=2.0.2",
    "google-search-results>=2.4.2",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
]
